+++ CHANGELOG.md
@@ -54,6 +54,8 @@ Upgrade nginz (#1658)
 
 ## Documentation
 
+* Improved Swagger documentation for endpoints with multiple responses (#1649).
+
 ## Internal changes
 
 * Improvements to local integration test setup when using buildah and kind (#1667)
@@ -104,6 +106,7 @@ This release requires a manual change in your galley configuration: `settings.co
 * [Federation] Adjust scripts under ./hack/federation to work with recent changes to the federation API (#1632).
 * Refactored Proteus endpoint to work with qualified users (#1634).
 * Refactored Federator InternalServer (#1637)
+* Introduced `MultiVerb` endpoints in Servant API (#1649).
 
 ### Internal Federation API changes